+++ todo-list-nextjs/package.json
@@ -19,8 +19,7 @@
     "@types/node": "20.12.7",
     "@types/react": "18.2.79",
     "eslint": "^8",
-    "eslint-config-next": "14.1.3",
-    "typescript": "5.4.5"
+    "eslint-config-next": "14.1.3"
   },
   "packageManager": "yarn@1.22.22+sha1.ac34549e6aa8e7ead463a7407e1c7390f61a6610"
 }
