+++ todo-list-nextjs/src/app/api/constants.ts
@@ -1,0 +1,31 @@
+const todos = [
+    {
+      id: 1,
+      title: "Todo 1",
+      createAt: "2024-03-29 23:39",
+      deadline: "2024-03-29 23:39",
+      description: "Todo 1 description",
+      completed: false,
+      status: "new"
+    },
+    {
+      id: 2,
+      title: "Todo 2",
+      createAt: "2024-03-29 23:39",
+      deadline: "2024-03-29 23:39",
+      description: "Todo 2 description",
+      completed: false,
+      status: "new"
+    },
+    {
+      id: 3,
+      title: "Todo 3",
+      createAt: "2024-03-29 23:39",
+      deadline: "2024-03-29 23:39",
+      description: "Todo 3 description",
+      completed: false,
+      status: "new"
+    },
+  ];
+
+export default todos
